This CL changes our base IDE in Android Studio from
15.0.3 (tag idea/143.1821.5) to
16.1.1 (tag idea/145.597.3)

+++ plugins/IntentionPowerPak/src/com/siyeh/ipp/exceptions/ConvertCatchToThrowsIntention.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2007-2014 Bas Leijdekkers
+ * Copyright 2007-2016 Bas Leijdekkers
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -20,6 +20,7 @@ import com.intellij.openapi.project.Project;
 import com.intellij.psi.*;
 import com.intellij.psi.util.PsiTreeUtil;
 import com.intellij.util.IncorrectOperationException;
+import com.siyeh.ig.psiutils.VariableSearchUtils;
 import com.siyeh.ipp.base.Intention;
 import com.siyeh.ipp.base.PsiElementPredicate;
 import org.jetbrains.annotations.NotNull;
@@ -61,7 +62,7 @@ public class ConvertCatchToThrowsIntention extends Intention {
     addToThrowsList(throwsList, catchType);
     final PsiTryStatement tryStatement = catchSection.getTryStatement();
     final PsiCatchSection[] catchSections = tryStatement.getCatchSections();
-    if (catchSections.length > 1 || tryStatement.getResourceList() != null) {
+    if (catchSections.length > 1 || tryStatement.getResourceList() != null || tryStatement.getFinallyBlock() != null) {
       catchSection.delete();
     }
     else {
@@ -69,12 +70,18 @@ public class ConvertCatchToThrowsIntention extends Intention {
       if (tryBlock == null) {
         return;
       }
-      final PsiElement first = tryBlock.getFirstBodyElement();
-      final PsiElement last = tryBlock.getLastBodyElement();
-      if (first != null && last != null) {
-        tryStatement.getParent().addRangeAfter(first, last, tryStatement);
+      final PsiCodeBlock parentCodeBlock = PsiTreeUtil.getParentOfType(tryStatement, PsiCodeBlock.class);
+      if (parentCodeBlock == null || !VariableSearchUtils.containsConflictingDeclarations(tryBlock, parentCodeBlock)) {
+        final PsiElement first = tryBlock.getFirstBodyElement();
+        final PsiElement last = tryBlock.getLastBodyElement();
+        if (first != null && last != null) {
+          tryStatement.getParent().addRangeAfter(first, last, tryStatement);
+        }
+        tryStatement.delete();
       }
-      tryStatement.delete();
+      else {
+        tryStatement.replace(tryBlock);
+      }
     }
   }
